+++ src/onegov.ballot/HISTORY.rst
@@ -11,6 +11,9 @@ Changelog
 - Fixes spelling of "eligible voters".
   [msom]
 
+- Adds election composites.
+  [msom]
+
 2.0.0 (2018-01-23)
 ~~~~~~~~~~~~~~~~~~~